requirements.txt
Django==2.2.6
django-ckeditor==5.7.1
django-debug-toolbar==2.0
django-filter==2.2.0
django-js-asset==1.2.2
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
env-file==2019.4.13
Markdown==3.1.1
psycopg2==2.8.3
public==2019.4.13
PyJWT==1.7.1
python-decouple==3.1
pytz==2019.2
sqlparse==0.3.0
values==2019.4.13
